usb: musb: cppi41: sw workaround fix for extra in token dma issue(2)
Fixes extra IN token issue seen when a file is transferred from one pendrive
to other. The same issue is also seen when a file is read from a usb pendrive
and same time an usb wifi module is connected and its ip interface is up.

Issue root caused to be with xDMA corrupting the data toggle bit when multiple
bus masters are accessing the mentor core register space.

Signed-off-by: Ajay Kumar Gupta <ajay.gupta@ti.com>
Signed-off-by: Ravi B <ravibabu@ti.com>