drivers: net: cpsw: Fix event loss and reduce lookup times in cpts_time_evts_fifo_pop()
A bug caused the current head item to be lost whenever
cpts_time_evts_fifo_pop() returned an item (!= head) inside the FIFO.
Another bug prevented the FIFO's item count to decrease once it reached
its maximum size, which increased event lookup times.
NOTE: The cpts_time_evts_fifo_*() functions and datastructure names
are misleading because items which are not at the head of the FIFO
can be removed.

Signed-off-by: Delio Brignoli <dbrignoli@audioscience.com>
Signed-off-by: Mugunthan V N <mugunthanvnm@ti.com>