mtd: nand: omap2: Fix bit flip in data area due to bit flip in OOB
Bit flips can be reported in OOB area and BCH8 ecc is capable of
recovering those bit flips. But due to the bug in driver, bit flips
errors reported in OOB area is going to correct in data area of the next
sector, which will cause data mismatch. This patch fixes the same by
correcting the bit flips in read_ecc.

Signed-off-by: Philip, Avinash <avinashphilip@ti.com>