ASoC: Fix swapping of channels after XRUN of stereo capture
Capturing 16 bit format audio leads to channel swap after an overrun.

After XRUN condition the driver needs to cleanup any residual data.

The cleanup process is initiated from the PCM layer in the following
order where the DMA is stopped before the CPU-DAI. This may cause the
CPU-DAI to generate a request for transfer another 2 byte (one sample)
of data which will remain in the queue and served when the DMA is
resumed.
This leads to the out of order operation of the CPU-DAI and DMA will
result in a channel swap.

Modifying the sequence in the following order will prevent the above
race condition and hence solves the channel swap issue.

START:	CODEC > PLATFORM > CPU-DAI
STOP:	CODEC > CPU-DAI > PLATFORM

This patch is verified on TI DM8168 EVM using AIC3106 CODEC and McASP
interface.

Signed-off-by: Raj, Deepu <deepu.raj@ti.com>