ASoC: Fix Davinci McASP audio loopback resume kernel crash
Kernel crashes while suspending to RAM due to audio loopback.
McASP functional clock is disabled during the suspend process.
The register access to the McASP module will result in a data
abort if the clock is already disabled by the capture/playback
instance in the case of the audio loopback.

Modifying the suspend function to check the clock state before
accessing the McASP register solved the kernel crash.

Signed-off-by: Raj, Deepu <deepu.raj@ti.com>