ti814x: PM: Reset unused clocks causes suspend to hang
Enabling the "Reset unused clocks" kernel option causes
suspend to memory operation to hang. On TI814X EVM the mmc instance
populated is MMC2 unlike MMC1 on TI816X EVM. But the instance number was passed as 1
in the board file which enabled the clock for MMC1 and not MMC2.
MMC2 was working as the clock was enabled in u-boot. When reset unused clock
was enabled, mmc2 clock was disabled and in suspend sequence mmc driver was found to
hang in clearing pending work.

This was fixed by correcting the mmc instance number and related base address
in board file.

Signed-off-by: Ravikumar Kattekola <rk@ti.com>
(cherry picked from commit 6c39f326fbd1ccdbc227557e78ccab682276960d)