mtd:nand:omap2: updated oobfree.offset for various ecc schemes
first 2 bytes of first page of every NAND block is reserved for
bad-block marker.
Then based on driver ECC.layout is decided based on number of ECC bytes
required. Remaining bytes are left free for storing file-system
metadata.
This patches fixes the oobfree.offset so that it does not overlaps with
bad-block marker offsets

Signed-off-by: Gupta, Pekon <pekon@ti.com>