ti81xx: Add clock domains for shared I2C clkctrl register.
The I2C busses on the ti814x share a common clkctrl register, and thus
we can't rely on the omap_hwmod enable/disable functions to know when
the other user of the register is active. Work around it by creating a
clock domain for the I2C modules which share clkctrl registers.